Rename SettingsActivity to LegacySettingsActivity
